requirements.txt
requests
flask
newsapi-python
pysqlite3
beautifulsoup4
